+++ src/StyleSheetManager/StyleSheetManager.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {{addStyleSheet: (style_sheet: CSSStyleSheet, beginning: boolean) => Promise<void>, generateVariableStyleSheet: (prefix: string, variables: {[key: string]: string}, beginning: boolean) => Promise<void>}} StyleSheetManager
+ */